Change EqualizerController to AudioController
Player is now an independent object decoupled (I think) from AudioController. Equalizer is also independent and decoupled from AudioController. Both inherit from the AVAudioNode subclasses that were being used previously. AudioController has instances of these classes as its nodes. The AudioController is essentially the music player backend for the the app UI to communicate with, providing a streamlined API to the view controllers and managing the deeper details internally